requirements.txt
Pillow==9.2.0
pygame==2.0.1
PyQt5==5.15.7
